SPDB-174: table_options: set the default filter policy to the SpeeDB one
If we don't set it, a change that was introduced upstream in commit
865a25101da248644169cbbe930980afb05adc4e will reset the filter policy
in db_bench and show worse results in our benchmarks. Unfortunately
this requires including "rocksdb/filter_policy.h" in "rocksdb/table.h"
in order to have a complete type of FilterPolicy for the instantiation
of the filter_policy member, which wasn't required before this change.

Note that this is a change that's incompatible with RocksDB, since it
expects the user to use our headers when compiling code that links with
SpeeDB, otherwise the filter policy can again be reset by db_bench (and
maybe other code paths that reach into the table factory after it has
been created).

While we're at it, fix the persisting and parsing of filter policies
and add support for creating them from existing DB options files.