SPDB-102: flush_job: don't merge more memtables than asked for
Merging memtables increases CPU usage and may create large SST files that
hinder compaction later on as it causes more overlap than intended, so we
should not merge more memtables than the amount allowed on the inactive
memtable list.

There are two types of flushes, regular and atomic, that we need to handle.

In the atomic flush case we need to create as many flush jobs as possible
for a given column family (bounded by max_memtable_id) and make sure to
execute them all.

In the non atomic flush case, since we aren't bounded by a max_memtable_id,
we just create flush jobs until there are no pending memtables to flush
or until we should stop (because the column family was dropped, an error
occurred, or a shutdown was requested).

Also adjust the relevant FlushJob and MemtableList tests.