SPDB-197: options: don't set memtable_whole_key_filtering in the presence of user comparators
This will create a memtable bloom filter if the user happens to also set
memtable_prefix_bloom_size_ratio to a value greater than 0.0, which is
incompatible with user comparators that don't consider the entire key
representation as is.

While we're at it, add a validation for filter policies in Block Based
table in conjunction with a user comparator that may consider different
byte contents being equal, and also check for memtable_whole_key_filtering.