SPDB-179: db_impl: add an experimental latency-optimised write flow
This write flow doesn't wait for the WAL to finish a write before writing
concurrently to the memtable, shortening the amount of time threads have
to wait for a write to complete in a heavily multi-threaded write scenario.

This commit includes code from commits for the following issues as well:
SPDB-269, SPDB-275, SPDB-278, SPDB-280, SPDB-281, SPDB-286, SPDB-291,
SPDB-292, SPDB-298, SPDB-310.