SPDB-243: test_util: fix custom comparators breakage after SPDB-197
SPDB-197 introduced a stricter check for custom comparators, and that
apparently broke test_util's custom comparators which don't override the
CanKeysWithDifferentByteContentsBeEqual() method, and which by default
returns true, which in turn causes these comparators to triggers the
stricter check even though they don't, in fact, consider keys with different
byte contents equal.

This is easily fixed by overriding this method to return false as it should.