SPDB-240: db_stress: change the way the expected values file is initialised
The current way of initialising the expected values file might cause an
error to be raised if the initialisation didn't finish by the time the
run was killed, as the file would still contain zeros which are interpreted
as valid value seeds instead of deletion sentinel.

Fix it by setting deletion sentinel to 0, so that after initialisation
we'll have all of the keys set to the needed value, and adapt the code
that generates values accordingly so that it doesn't end up using a sentinel
value as a seed.

While at it, try to avoid excess allocation by first trying to Truncate
in order to initialise the file, and move to Append with an allocated
zeros buffer only if doesn't achieve the expected result.