SPDB-334: ci: enforce branch naming and commit message conventions
This will reject any branch that is submitted for a merge request if its
name doesn't contain a JIRA key, or if any of its commits don't begin with
the branch's JIRA key.

While at it, refactor the CI config to use a smaller (and version-pinned)
base image and clean up the existing job definitions.