SPDB-355: compaction_job: don't break a file if not on key boundary
If we have files that are broken in such a key that a key appears in both
of them (e.g. due to snapshots being held, so that the two keys have different
sequence numbers), if we pick only the first of these two files, we'll
end up causing the older key (with the smaller sequence number) becoming
visible again as it'll stay in a higher level.

Similarly, if those two files are on the last level and we pick the first
one, after compaction its sequence number would be zeroed out, and it would
compare larger than the other one, causing an overlap and an error.

To solve these cases, prevent equal user keys from being split between
files even if it violates some other heuristic (such as maximum output
file size).