SPDB-228: hybrid_compaction: choose all versions of a key for level compaction
If two versions of a key end up in different files on the same level, we
need to take care to never choose only the first file, as it contains the
newer of the keys (newer versions compare lower than older ones), and a
compaction will move it to a lower level, violating the LSM invariant that
newer keys are always on the same level as older keys or on a higher one.
Choosing only the second file is OK in this case, though, as it only moves
the older key to a lower level.

Similarly, when choosing files from the bottom-most level, we cannot choose
the first of the two files, as due to sequence number zeroing the newer
key will now have a smaller sequence number than the older one in the second
file and that would cause an overlap in the bottom-most level.

To ensure this doesn't happen, choose all of the versions of a key in the
target level and in the lowest level of the hyper-level being compacted.
On upper levels in the hyper-level we just exclude such files in order
to maintain the lower- and upper-bound restrictions.