SPDB-470: external_delay: only print a reset message when we actually reset
After SPDB-219, the log is bombarded with useless "Resettign external delay"
messages because it's printed whenever we calculate a zero delay, even if
a zero delay is what's already set. This hurts performance and fills the log
with garbage.

Change the ExternalDelay interface to report if it actually performed a reset
and only print the message when it actually does.