SPDB-496: external_delay: avoid signed integer overflow
UBSAN reported a signed integer overflow on the subtraction inside the
std::min() call due to the conversion to an int. Use proper std::chrono
types to make the code more self explanatory regarding units used and have
higher precision. We still don't expect to wait more than 35 minutes per
write (which is the maximum amount of microseconds that a signed int is
able to represent) since we're limiting the sleep time to a maximum of a
second, so the conversion to in the call to clock->SleepForMicroseconds()
is OK.