SPDB-544: ci: add lint checks for Python and shell code
This uses lint-diffs to only check the diff and output errors from
multiple linters.

For Python we use flake8 and for shell code we use shellcheck. However,
we can't use clang-tidy without generating a compilation database, which
the linting stage in the pipeline shouldn't do (as it requires cmake and
a compiler installed) and format-diff.sh doesn't generate output that can
be consumed by lint-diffs, so this commit instead adds a wrapper script
around clang-format so we can continue to lint only code style violations.