SPDB-550: ci: set flake8 line length limit to 88 to match Black
Black's default is 88 characters, and it would be easier if people didn't
need to fix up after it just because flake8 is more conservative.

This still doesn't solve string literals, which Black doesn't currently
break, but hopefully these are rather rare cases that would not take up
too much time and effort to manually fix.