SPDB-555: ci: configure a cache for all builds
Currently we build everything as a sanity check to ensure that broken code
doesn't get committed into master. This takes a long time, and there's
no reason to repeat it on every MR if we only want to know that the code
successfully builds.

Add ccache and configure a runner cache for the build verification job (for
now let's use the ccache default of 5GB. It should suffice for the current
size of the code. We might want to change that in the future by setting
the CCACHE_MAXSIZE environment variable).