WBM: Fix stall deadlock with multiple cfs
With a setting of multiple cfs and WriteBufferManager with allow_stall,
the DB can enter a deadlock when the WBM initiates a stall.
This happens since only the oldest cf is picked for flush when
HandleWriteBufferManagerFlush is called to flush the data and prevent the stall.
When using multiple CFs, this does not ensure the FreeMem will evict
enough memory to prevent a stall and no other flush is scheduled.

To fix this, add cfs to the flush queue so that we'll be below the mutable_limit_.